Command Pattern has been deleted, changing the Result to GameMode and added two functions, every chess now have a rank which is ranked by their height. isKingCaptured function, endGame function, getRank function......etc has been added. As there is no surrender function and no test cases have been created. If you want to check the result, you either need to finish the whole chess game or wait for me to release the draft of the getResult test case. If there is any questions or any requests, please directly post on the wtsapp group and we can discuss later, thank you for your cooperation!

+++ ChessGame/src/chessGame/ChessPlayer.java
@@ -14,7 +14,7 @@ public class ChessPlayer {
 		chessPiece=16;
 	}
 	
-	public void addScore(int score){
+	public void addScore(double score){
 		playerScore+=score;
 	}
